requirements.txt
# Python Dependencies untuk Data Preprocessing & Supabase Integration

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # untuk Excel files

# Supabase Integration
supabase>=2.0.0
python-dotenv>=1.0.0

# Optional: Data Analysis & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Optional: Advanced Analytics
scikit-learn>=1.3.0  # untuk machine learning
